requirements.txt
requests
pyodbc
python-dotenv
pandas
deep-translator
cryptography